Credit Card Fraud Detection
Credit card fraud detection involves using specific tools and tactics to recognize a potentially fraudulent purchase. This means that the person placing the order is not the rightful owner of the credit card.

For online businesses, there are tools that analyze transactions for risk of fraud, based on verifying CVV, billing address, IP address, and other matches.

For face-to-face purchases, fraud detection will require the staff's diligence on observing behavior, verifying photo ID, and examining the credit card.

We'll go over these different methods in-depth, for both online and in-store.