WebsiteLinks
This internship is about the techniques which are used to fetch the details and provide data prediction & data learning with some datasets and can be further updated. I'm using python language for prediction the details in a serialized manner ,datasets- Credit Card, Iris-Data, Movie-Prediction. Do visit the website for further process you can message me on my linkedIn profile.